backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pymongo==4.3.3
python-dotenv==1.1.0
pytest==7.4.0
